requirements.txt
customtkinter==5.2.2
pillow==10.4.0
requests==2.32.3
matplotlib==3.9.2
pandas==2.2.3
tkcalendar==1.6.1